+++ modules/account/index.ts
@@ -1,5 +1,10 @@
 export { getAccountLinksEnabled, useAccountLinksEnabled } from "./core/accountLinksConfig";
 export { getAccountCabinetUrl, openAccountCabinet, type CabinetContext } from "./core/openAccountCabinet";
-export { resolveBillingCurrency, type BillingCurrency } from "./core/billingCurrency";
+export {
+  resolveBillingCurrency,
+  resolveBillingGeo,
+  type BillingCurrency,
+  type BillingGeo,
+} from "./core/billingCurrency";
 export { deleteAccountRemote } from "./core/deleteAccountClient";
 export { MembershipEventsBridge } from "./ui/MembershipEventsBridge";